Enhanced to slide a head horizontally.
1. This patch includes addition of TextViews for title and time. But
TextView for time is not used. Because updating it causes re-layouting.
The head returns to top-left corner on re-layouting.

The uzumaki slider has responsibility to layout the head. So the head
must be a child of the slider. Patch in future will fix them.

2. After releasing the head, the head sometimes moves to most out
position. This is a bug. The reason of it is unknown still.