Fixed the bug by which background of some list items becomes black.
I changed to use setBackgroundDrawable() with ColorDrawable in the
past commit, but it is not enough. Some list's rows show strange
background color. I guessed that this is also a bug of Android similaly
to View.setBackgroundColor(). So I decided to change screen design for
unused background color.

In new design, the middle pane has one TextView on the top of the file
list. It shows the directory path of the file list.