Fixed to work on Android 4.2.2.
The main problem is that android.media.MediaPlayer in Android 4.2.2 is
more strict than older one. For old Android, AAP called
MediaPlayer.reset() in any cases with ignoring the status of a
MediaPlayer object. It worked well. But MediaPlayer in Android 4.2.2
does not work for reset() to seek. MediaPlayer requires to call seekTo()
when AAP seeks. Calling seekTo() is needed even when music has been
paused. So now we must know which music a MediaPlayer object handles.

The new version of AAP does not repeat starting and stopping the
service. The service is alive always to memorize the current music.

The bug happened only in playing long music like Internet radio.